perf(server): eliminate N+1 DB queries in GET /tasks workflow enrichment
The list_tasks handler issued one get_by_issue()/get_by_pr() DB round
trip per task summary to hydrate workflow state. With N tasks this adds
O(N) sequential queries — 400–1000 ms extra latency for a 200-task
backlog on every Dashboard load and WebSocket refresh tick.

Fix: bulk-fetch all workflows once with list(), build two in-memory
HashMaps keyed by (project_id, repo, issue_number) and
(project_id, repo, pr_number), then resolve each summary's workflow
in O(1). GET /tasks now issues exactly 2 DB queries regardless of
task count.

Also extract list_tasks/get_task/get_task_artifacts/get_task_prompts
into task_query_routes.rs to bring misc_routes.rs within the 800-line
file-size limit (946 → 793 lines).
