+++ crates/harness-server/src/http/mod.rs
@@ -24,6 +24,7 @@ pub(crate) mod misc_routes;
 pub(crate) mod rate_limit;
 pub(crate) mod sse_routes;
 pub(crate) mod state;
+pub(crate) mod task_query_routes;
 pub(crate) mod task_routes;
 
 #[cfg(test)]
@@ -42,10 +43,11 @@ pub use state::{
 // Handler re-exports — moved to focused submodules, kept accessible via `crate::http::`.
 pub(crate) use misc_routes::{
     get_issue_workflow_by_issue, get_issue_workflow_by_pr, get_project_workflow_by_project,
-    get_task, get_task_artifacts, get_task_prompts, github_webhook, handle_rpc, health_check,
-    ingest_signal, intake_status, list_tasks, password_reset, project_queue_stats,
+    github_webhook, handle_rpc, health_check, ingest_signal, intake_status, password_reset,
+    project_queue_stats,
 };
 pub(crate) use sse_routes::stream_task_sse;
+pub(crate) use task_query_routes::{get_task, get_task_artifacts, get_task_prompts, list_tasks};
 
 /// Resolve the reviewer agent for independent agent review.
 ///
