+++ crates/harness-server/src/http/task_query_routes.rs
@@ -1,0 +1,179 @@
+use axum::{
+    extract::{Path, State},
+    http::StatusCode,
+    response::{IntoResponse, Response},
+    Json,
+};
+use serde_json::json;
+use std::collections::HashMap;
+use std::sync::Arc;
+
+use super::state::AppState;
+
+pub(crate) async fn list_tasks(State(state): State<Arc<AppState>>) -> Response {
+    match state.core.tasks.list_all_summaries_with_terminal().await {
+        Ok(mut summaries) => {
+            if let Some(workflow_store) = state.core.issue_workflow_store.as_ref() {
+                // Bulk-fetch all workflows once to avoid O(N) sequential DB round trips.
+                let workflows = workflow_store.list().await.unwrap_or_else(|e| {
+                    tracing::error!("list_tasks: failed to bulk-fetch workflows: {e}");
+                    Vec::new()
+                });
+                let mut workflows_by_issue = HashMap::new();
+                let mut workflows_by_pr = HashMap::new();
+                for wf in workflows {
+                    let issue_key = (wf.project_id.clone(), wf.repo.clone(), wf.issue_number);
+                    if let Some(pr_num) = wf.pr_number {
+                        let pr_key = (wf.project_id.clone(), wf.repo.clone(), pr_num);
+                        workflows_by_issue
+                            .entry(issue_key)
+                            .or_insert_with(|| wf.clone());
+                        workflows_by_pr.entry(pr_key).or_insert(wf);
+                    } else {
+                        workflows_by_issue.entry(issue_key).or_insert(wf);
+                    }
+                }
+                for summary in &mut summaries {
+                    let Some(project_id) = summary.project.as_deref() else {
+                        continue;
+                    };
+                    let by_issue = summary
+                        .external_id
+                        .as_deref()
+                        .and_then(|id| id.strip_prefix("issue:"))
+                        .and_then(|n| n.parse::<u64>().ok());
+                    let by_pr = summary
+                        .external_id
+                        .as_deref()
+                        .and_then(|id| id.strip_prefix("pr:"))
+                        .and_then(|n| n.parse::<u64>().ok())
+                        .or_else(|| {
+                            summary
+                                .pr_url
+                                .as_deref()
+                                .and_then(crate::http::parse_pr_num_from_url)
+                        });
+                    summary.workflow = match (by_issue, by_pr) {
+                        (Some(issue), _) => workflows_by_issue
+                            .get(&(project_id.to_owned(), summary.repo.clone(), issue))
+                            .cloned(),
+                        (None, Some(pr)) => workflows_by_pr
+                            .get(&(project_id.to_owned(), summary.repo.clone(), pr))
+                            .cloned(),
+                        (None, None) => None,
+                    };
+                }
+            }
+            Json(summaries).into_response()
+        }
+        Err(e) => {
+            tracing::error!("list_tasks: database error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+pub(crate) async fn get_task(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    match state
+        .core
+        .tasks
+        .get_with_db_fallback(&harness_core::types::TaskId(id))
+        .await
+    {
+        Ok(Some(task)) => Json(task).into_response(),
+        Ok(None) => (
+            StatusCode::NOT_FOUND,
+            Json(json!({"error": "task not found"})),
+        )
+            .into_response(),
+        Err(e) => {
+            tracing::error!("get_task: database error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /tasks/{id}/artifacts — all persisted artifacts for a task.
+pub(crate) async fn get_task_artifacts(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    let task_id = harness_core::types::TaskId(id);
+    match state.core.tasks.get_with_db_fallback(&task_id).await {
+        Ok(None) => {
+            return (
+                StatusCode::NOT_FOUND,
+                Json(json!({"error": "task not found"})),
+            )
+                .into_response();
+        }
+        Err(e) => {
+            tracing::error!("get_task_artifacts: database error: {e}");
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response();
+        }
+        Ok(Some(_)) => {}
+    }
+    match state.core.tasks.list_artifacts(&task_id).await {
+        Ok(artifacts) => Json(artifacts).into_response(),
+        Err(e) => {
+            tracing::error!("get_task_artifacts: list artifacts error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /tasks/{id}/prompts — all persisted redacted prompts for a task.
+pub(crate) async fn get_task_prompts(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    let task_id = harness_core::types::TaskId(id);
+    match state.core.tasks.get_with_db_fallback(&task_id).await {
+        Ok(None) => {
+            return (
+                StatusCode::NOT_FOUND,
+                Json(json!({"error": "task not found"})),
+            )
+                .into_response();
+        }
+        Err(e) => {
+            tracing::error!("get_task_prompts: database error: {e}");
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response();
+        }
+        Ok(Some(_)) => {}
+    }
+    match state.core.tasks.get_prompts(&task_id).await {
+        Ok(prompts) => Json(prompts).into_response(),
+        Err(e) => {
+            tracing::error!("get_task_prompts: query error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
